+++ Changelog.md
@@ -76,6 +76,8 @@
 - Show slicing speed in layers/sec
 - Support highlighting selected mesh sections
 - Hide Auto Layout annealing settings by default
+- Calculate mesh volume asynchronously (helps on models with ~ >10M faces)
+- Reduce unnecessary cross-thread memory copies
 
 ## v0.9.2 &mdash; August 30th, 2026